pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agldm"
version = "0.1.0"
description = "Attribute-guided zero-shot text-to-image synthesis research MVP"
requires-python = ">=3.10"
dependencies = [
  "lpips",
  "numpy",
  "pillow",
  "pyyaml",
  "scipy",
  "torch",
  "torchvision",
  "transformers",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
agldm = "agldm.cli:main"

[tool.setuptools.packages.find]
include = ["agldm*"]
